Only run file generator if paths match each other (failed)
This was an attempt at fixing this issue:
https://github.com/Dan503/gulp-file-loader/issues/3

I needed to do most of this work as part of fixing this issue anyway:
https://github.com/Dan503/gulp-file-loader/issues/1

I successfully managed to get the compiler to only run if the paths matched, but then I completely changed the format of the output and it didn't change the format of the output because the paths were still all the same. >_<

I think I will just close issue #3. The only way I can make sure it works is if I go almost all the way to creating the file but stop just short of outputting something.