Adding dest to the output message
I had a moment of "where did the file go? I can't find it" and so I'm sure that other users are going to run into that issue as well. I've added the destination folder to the file generation confirmation message so that users are able to track down missing files more easily.

This also makes it easier to debug when the value of "dest" isn't what it should be.