Replacing the old "other settings" section with a full "settings reference guide"
I realised that I didn't really have a good place for people to go if they just need a quick refresher on what a setting is for. I needed somewhere that listed out all the settings and explained what they did.

I'm also adding the documentation required for issue #4 in this commit.
https://github.com/Dan503/gulp-file-loader/issues/4