Using an arrow function instead of the word "function"
It is a bit more ES6 flavoured and uses less characters.

It isn't as human readable though. It's a bit easier to understand what it is doing if you explicitly say "function"

I'm going to revert this commit. I want it in my git hystory though incase I change my mind later.